Update Pentacrab.py
I added a new attack. And I cannot beat the boss anymore. I usually try to beat the boss before I update this thing. Yet I couldn't beat it. It is too powerful. I will probably add a shield ability in order to allow the player to not get evicerated within an instant.