Update Pentacrab.py
Its done. All my hard work ending today. I have finished Pentacrab. If it is good or not is up for debate. But its done. I added a nice visualizer for the shield health and buffed the laser and dive attack. I also fixed some of the sound issues caused by the near constant spam of fireballs near the end of the boss. Its still aint the best but its better that how it was before. Good luck for anyone who tried to beat this.